backend/requirements.txt
fastapi==0.115.2
passlib==1.7.4
pydantic==2.9.2
PyJWT==2.9.0
python-decouple==3.8
psycopg2-binary
pydantic[email]
email-validator
SQLAlchemy==2.0.35
uvicorn==0.32.0
